[refactor] API에서 데이터 받아오는 NetworkManager 클래스와 코어데이터 CRUD 작업하는 클래스 PokeDataManager 클래스 분리
